+++ NEWS.md
@@ -40,4 +40,7 @@
 
 # odytools 0.4.0
 
--   Adding AEs functions... (branch aes)
+-   Added AE related functions.
+
+    -   **ody_make_ae_tbl**: It makes a "max grede" count of patients.
+    -   **ody_make_ae_gt**: It inputs the output of ody_make_ae_tbl and makes a gt display table.