+++ NEWS.md
@@ -42,5 +42,5 @@
 
 -   Added AE related functions.
 
-    -   **ody_make_ae_tbl**: It makes a "max grede" count of patients.
+    -   **ody_make_ae_tbl**: It makes a "max grade" count of patients.
     -   **ody_make_ae_gt**: It inputs the output of ody_make_ae_tbl and makes a gt display table.